User can see paginated list of documents
Summary:
Rework document list page design [based on mockups](https://opengovfoundation.github.io/madison-mockups/documents.html).

Also include design updates for M5 (T267, and T268)

Misc:
* Use `introtext_html` in document show view.
* Remove tests for deleting documents from list view.
* Various code cleanup.

Resolves T248
Resolves T267
Resolves T268

Test Plan:
* Check the document list page against [the mockups](https://opengovfoundation.github.io/madison-mockups/documents.html)
* Check the design against M5
* Check the homepage for the updated document card design

Reviewers: doshitan

Reviewed By: doshitan

Maniphest Tasks: T248, T267, T268

Differential Revision: https://phabricator.opengovfoundation.org/D151

+++ app/Policies/DocumentPolicy.php
@@ -92,6 +92,9 @@ class DocumentPolicy extends Policy
 
     public function restore(User $user, Document $document)
     {
+        if ($document->publish_state === Document::PUBLISH_STATE_DELETED_ADMIN) {
+            return $user->isAdmin();
+        }
         return $this->delete($user, $document);
     }
